TALKING DATA: Train/Test Data with 72 variables
This file converges what Jordan did with my own cleaning that was required for the recommender system.  I was unable to add all the variables using his original file, so I had to get creative.  I think running all this code would take a full day, so I am uploading the cleaned train/test datasets to the Google Drive.  It has all of my 72 variables, but they have not/not been tested for predictive power nor the recommender system.  I had to center and scale the data for those purposes.